+++ app/src/main/java/com/nimontoy/android/controller/activity/login/LoginActivity.kt
@@ -35,7 +35,7 @@ class LoginActivity : BaseActivity() {
     private lateinit var callbackManager: CallbackManager
     private lateinit var auth : FirebaseAuth
 
-    private lateinit var googleSignInClient: GoogleSignInClient
+    private val googleLogin = GoogleLogin(this)
     //kakao
     //private var callback : SessionCallback? = null
 
@@ -102,7 +102,7 @@ class LoginActivity : BaseActivity() {
             try {
                 // Google Sign In was successful, authenticate with Firebase
                 val account = task.getResult(ApiException::class.java)
-                firebaseAuthWithGoogle(account!!)
+                googleLogin.firebaseAuthWithGoogle(account!!, auth)
                 println (account)
             } catch (e: ApiException) {
                 // Google Sign In failed, update UI appropriately
@@ -137,7 +137,7 @@ class LoginActivity : BaseActivity() {
             }
     }
 
-    private fun updateUI(user: FirebaseUser?) {
+    fun updateUI(user: FirebaseUser?) {
         if (user != null) {
             /*
             status.text = getString(R.string.facebook_status_fmt, user.displayName)
@@ -179,65 +179,15 @@ class LoginActivity : BaseActivity() {
     }
 */
 
-
     companion object {
-        private const val RC_SIGN_IN = 9001
+        const val RC_SIGN_IN = 9001
     }
 
-    // [START auth_with_google]
-    private fun firebaseAuthWithGoogle(acct: GoogleSignInAccount) {
-        Log.d(TAG, "firebaseAuthWithGoogle:" + acct.id!!)
 
-        val credential = GoogleAuthProvider.getCredential(acct.idToken, null)
-        auth.signInWithCredential(credential)
-            .addOnCompleteListener(this) { task ->
-                if (task.isSuccessful) {
-                    // Sign in success, update UI with the signed-in user's information
-                    Log.d(TAG, "signInWithCredential:success")
-                    val user = auth.currentUser
-                    user?.let {
-                        // Name, email address, and profile photo Url
-                        val name = user.displayName
-                        val email = user.email
-                        val photoUrl = user.photoUrl
-
-                        // Check if user's email is verified
-                        val emailVerified = user.isEmailVerified
-
-                        // The user's ID, unique to the Firebase project. Do NOT use this value to
-                        // authenticate with your backend server, if you have one. Use
-                        // FirebaseUser.getToken() instead.
-                        val uid = user.uid
-                        println (name)
-                        println (email)
-                        println (photoUrl)
-                        println (emailVerified)
-                        println (uid)
-                    }
-                    updateUI(user)
-                } else {
-                    // If sign in fails, display a message to the user.
-                    Log.w(TAG, "signInWithCredential:failure", task.exception)
-                    updateUI(null)
-                }
-            }
-    }
-    // [END auth_with_google]
 
     private fun googleLogin () {
-        val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-            .requestIdToken(getString(R.string.default_web_client_id))
-            .requestEmail()
-            .build()
-
-        googleSignInClient = GoogleSignIn.getClient(this, gso)
         google_login.setOnClickListener {
-            googleSignIn()
+            googleLogin.login()
         }
-    }
-
-    private fun googleSignIn() {
-        val signInIntent = googleSignInClient.signInIntent
-        startActivityForResult(signInIntent, RC_SIGN_IN)
     }
 }
